fix: attendance atomic function confidence_score type mismatch
Fixed PostgreSQL error "invalid input syntax for type numeric: 'high'"
caused by incorrect cast in reprocess_prickle_attendance_atomic function.

The confidence_score column is text with CHECK constraint ('high', 'medium',
'low'), not numeric. Removed incorrect ::numeric cast.

Also updated export/restore scripts to use grep -oP for parsing DB URL
from newer Supabase CLI output format.

Changes:
- supabase/migrations/20260423000007: Remove ::numeric cast
- scripts/export-local-tables.sh: Fix DB URL parsing
- scripts/restore-local-tables.sh: Fix DB URL parsing, add note about FK errors

+++ supabase/migrations/20260423000007_add_reprocess_attendance_atomic.sql
@@ -59,7 +59,7 @@ BEGIN
     ),
     (attendance.value->>'join_time')::timestamptz,
     (attendance.value->>'leave_time')::timestamptz,
-    (attendance.value->>'confidence_score')::numeric
+    attendance.value->>'confidence_score'
   FROM jsonb_array_elements(new_attendance_data) AS attendance
   LEFT JOIN prickle_id_map ON prickle_id_map.client_prickle_id = attendance.value->>'client_prickle_id'
   WHERE new_attendance_data != 'null'::jsonb AND jsonb_array_length(new_attendance_data) > 0;
